Locking approach modifications. Special keys to control rotation.
1. Special keys
Special keys handler was added to control rotation of the cube. The actual
effects of arrows is still highly experimental. To be refined.

2. Locking.
Different lock system for image/texture updates has been implemented. Now all
the work for texture update is performed in timer callback. That texture
update and image update are wrapped by semaphors. When texture is updated then
texture is disabled for opengl.